Change D2D1PixelShader input APIs to use int values

+++ src/ComputeSharp.D2D1/Shaders/Interop/D2D1PixelShader.cs
@@ -85,12 +85,12 @@ public static class D2D1PixelShader
     /// </summary>
     /// <typeparam name="T">The type of D2D1 pixel shader to get the input count for.</typeparam>
     /// <returns>The number of inputs for the D2D1 pixel shader of type <typeparamref name="T"/>.</returns>
-    public static uint GetInputCount<T>()
+    public static int GetInputCount<T>()
         where T : unmanaged, ID2D1PixelShader
     {
         Unsafe.SkipInit(out T shader);
 
-        return shader.GetInputCount();
+        return (int)shader.GetInputCount();
     }
 
     /// <summary>
@@ -100,17 +100,17 @@ public static class D2D1PixelShader
     /// <param name="index">The index of the input to get the type for.</param>
     /// <returns>The type of the input of the target D2D1 pixel shader at the specified index.</returns>
     /// <exception cref="ArgumentOutOfRangeException">Thrown if <paramref name="index"/> is not in range for the available inputs for the shader type.</exception>
-    public static D2D1PixelShaderInputType GetInputType<T>(uint index)
+    public static D2D1PixelShaderInputType GetInputType<T>(int index)
         where T : unmanaged, ID2D1PixelShader
     {
         Unsafe.SkipInit(out T shader);
 
-        if (index >= shader.GetInputCount())
+        if ((uint)index >= shader.GetInputCount())
         {
             ThrowHelper.ThrowArgumentOutOfRangeException(nameof(index), "The input index is outside of range for the target pixel shader type.");
         }
 
-        return (D2D1PixelShaderInputType)shader.GetInputType(index);
+        return (D2D1PixelShaderInputType)shader.GetInputType((uint)index);
     }
 
     /// <summary>
